Sylvain, if faudrait que tu verifies...
La variable keyBXLevel doit avoir le meme type flottant que ProtMatchLevel
car elles sont comparees (avec >=) et deux reels identiques peuvent
avoir des representations differentes en float et en double.

+++ eugene/src/SensorPlugins/BlastX/Sensor.BlastX.h
@@ -15,7 +15,7 @@ class SensorBlastX : public Sensor
   std::vector<REAL> vPMLevel, vPMatch;
   std::vector<int>::iterator iter;
   int    index;
-  double keyBXLevel[10];
+  REAL keyBXLevel[10];
   int    minIn;
   int blastxM;
